+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gopjrt Changelog
 
+# v0.9.1 2025/11/07: More multi-device support
+
+- Package `pjrt`:
+  - Added `Buffer.CopyToDevice()`.
+
 # v0.9.0 2025/11/06: Multi-device SPMD execution; TPU installation support; Deprecating `xlabuilder`.
 
 - `gopjrt_installer`: Added TPU installation.